+++ pkgs/unified_analytics/lib/src/user_property.dart
@@ -5,13 +5,14 @@
 import 'dart:convert';
 
 import 'package:clock/clock.dart';
+import 'package:file/file.dart';
 
 import 'constants.dart';
-import 'session.dart';
+import 'event.dart';
+import 'initializer.dart';
 import 'utils.dart';
 
 class UserProperty {
-  final Session session;
   final String? flutterChannel;
   final String host;
   final String? flutterVersion;
@@ -22,11 +23,18 @@ class UserProperty {
   final String? clientIde;
   final String? enabledFeatures;
 
+  final File sessionFile;
+
+  /// Contains instances of [Event.analyticsException] that were encountered
+  /// during a workflow and will be sent to GA4 for collection.
+  final Set<Event> errorSet = {};
+
+  int? _sessionId;
+
   /// This class is intended to capture all of the user's
   /// metadata when the class gets initialized as well as collecting
   /// session data to send in the json payload to Google Analytics.
   UserProperty({
-    required this.session,
     required this.flutterChannel,
     required this.host,
     required this.flutterVersion,
@@ -36,13 +44,45 @@ class UserProperty {
     required this.locale,
     required this.clientIde,
     required this.enabledFeatures,
+    required this.sessionFile,
   });
+
+  /// This will use the data parsed from the
+  /// session file in the dart-tool directory
+  /// to get the session id if the last ping was within
+  /// [kSessionDurationMinutes].
+  ///
+  /// If time since last ping exceeds the duration, then the file
+  /// will be updated with a new session id and that will be returned.
+  ///
+  /// Note, the file will always be updated when calling this method
+  /// because the last ping variable will always need to be persisted.
+  int? getSessionId() {
+    _refreshSessionData();
+    final now = clock.now();
+
+    // Convert the epoch time from the last ping into datetime and check if we
+    // are within the kSessionDurationMinutes.
+    final lastPingDateTime = sessionFile.lastModifiedSync();
+    if (now.difference(lastPingDateTime).inMinutes > kSessionDurationMinutes) {
+      // Update the session file with the latest session id
+      _sessionId = now.millisecondsSinceEpoch;
+      writeSessionContents(sessionFile: sessionFile);
+    } else {
+      // Update the last modified timestamp with the current timestamp so that
+      // we can use it for the next _lastPing calculation
+      sessionFile.setLastModifiedSync(now);
+    }
+
+    return _sessionId;
+  }
 
   /// This method will take the data in this class and convert it into
   /// a Map that is suitable for the POST request schema.
   ///
-  /// This will call the [Session] object's [Session.getSessionId] method which
-  /// will update the session file and get a new session id if necessary.
+  /// This will call the [UserProperty] object's [UserProperty.getSessionId]
+  /// method which will update the session file and get a new session id
+  /// if necessary.
   ///
   /// https://developers.google.com/analytics/devguides/collection/protocol/ga4/user-properties?client_type=gtag
   Map<String, Map<String, Object?>> preparePayload() {
@@ -57,10 +97,56 @@ class UserProperty {
     return jsonEncode(_toMap());
   }
 
+  /// This will go to the session file within the dart-tool
+  /// directory and fetch the latest data from the session file to update
+  /// the class's variables. If the session file is malformed, a new
+  /// session file will be recreated.
+  ///
+  /// This allows the session data in this class to always be up
+  /// to date incase another tool is also calling this package and
+  /// making updates to the session file.
+  void _refreshSessionData() {
+    /// Using a nested function here to reduce verbosity
+    void parseContents() {
+      final sessionFileContents = sessionFile.readAsStringSync();
+      final sessionObj =
+          jsonDecode(sessionFileContents) as Map<String, Object?>;
+      _sessionId = sessionObj['session_id'] as int;
+    }
+
+    try {
+      // Failing to parse the contents will result in the current timestamp
+      // being used as the session id and will get used to recreate the file
+      parseContents();
+    } on FormatException catch (err) {
+      final now = createSessionFile(sessionFile: sessionFile);
+
+      errorSet.add(Event.analyticsException(
+        workflow: 'UserProperty._refreshSessionData',
+        error: err.runtimeType.toString(),
+        description: 'message: ${err.message}\nsource: ${err.source}',
+      ));
+
+      // Fallback to setting the session id as the current time
+      _sessionId = now.millisecondsSinceEpoch;
+    } on FileSystemException catch (err) {
+      final now = createSessionFile(sessionFile: sessionFile);
+
+      errorSet.add(Event.analyticsException(
+        workflow: 'UserProperty._refreshSessionData',
+        error: err.runtimeType.toString(),
+        description: err.osError?.toString(),
+      ));
+
+      // Fallback to setting the session id as the current time
+      _sessionId = now.millisecondsSinceEpoch;
+    }
+  }
+
   /// Convert the data stored in this class into a map while also
-  /// getting the latest session id using the [Session] class.
+  /// getting the latest session id using the [UserProperty] class.
   Map<String, Object?> _toMap() => <String, Object?>{
-        'session_id': session.getSessionId(),
+        'session_id': getSessionId(),
         'flutter_channel': flutterChannel,
         'host': host,
         'flutter_version': flutterVersion,